backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
openai==1.3.0
yt-dlp==2023.11.16
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
